requirements.txt
pytest==8.1.1
allure-pytest==2.13.4
requests==2.31.0